playwright_mcp/requirements.txt
pyqt5
python-dotenv
agent_framework
azure-identity
agent_framework_azure_ai
